+++ mne/filter.py
@@ -230,6 +230,15 @@ def _prep_for_filtering(x, copy, picks=None):
     x.shape = (np.prod(x.shape[:-1]), x.shape[-1])
     if picks is None:
         picks = np.arange(x.shape[0])
+    elif len(orig_shape) == 3:
+        n_epochs, n_channels, n_times = orig_shape
+        offset = np.repeat(np.arange(0, n_channels * n_epochs, n_channels),
+                           len(picks))
+        picks = np.tile(picks, n_epochs) + offset
+    elif len(orig_shape) > 3:
+        raise ValueError('picks argument is not supported for data with more'
+                         ' than three dimensions')
+
     return x, orig_shape, picks